Renamed entity file to EntityLayout
